move hwm_type validation in db_reader dialect class
