+++ onetl/hwm/__init__.py
@@ -1,2 +1,3 @@
 from onetl.hwm.auto_hwm import AutoHWM
+from onetl.hwm.decimal import DecimalHWM
 from onetl.hwm.statement import Statement
